+++ components/liquor/search/ranking/RankingKeywordSkeleton.tsx
@@ -1,23 +1,23 @@
 function RankingKeywordSkeleton() {
   return (
-    <div className="flex items-center gap-14">
-      <div className="flex flex-col">
+    <div className="flex items-center gap-x-[40px]">
+      <div className="flex flex-col gap-y-[13px]">
         {[...Array(5)].map((_, index) => (
-          <div key={index} className="my-1 flex items-center gap-3">
-            <div className="flex w-2.5 items-center justify-center text-sm font-bold text-suldak-gray-700">
+          <div key={index} className="flex items-center gap-x-3">
+            <div className="flex w-2.5 items-center justify-center text-[14px] font-bold text-suldak-gray-700">
               {index + 1}
             </div>
-            <div className="h-[18px] w-[118px] animate-pulse rounded-lg bg-suldak-gray-300" />
+            <div className="h-[17px] w-[80px] animate-pulse rounded-lg bg-suldak-gray-300" />
           </div>
         ))}
       </div>
-      <div className="flex flex-col">
+      <div className="flex flex-col gap-y-[13px]">
         {[...Array(5)].map((_, index) => (
-          <div key={index} className="my-1 flex items-center gap-3">
-            <div className="flex w-2.5 items-center justify-center text-sm font-bold text-suldak-gray-700">
+          <div key={index} className="flex items-center gap-x-3">
+            <div className="flex w-2.5 items-center justify-center text-[14px] font-bold text-suldak-gray-700">
               {index + 6}
             </div>
-            <div className="h-[18px] w-[118px] animate-pulse rounded-lg bg-suldak-gray-300" />
+            <div className="h-[17px] w-[80px] animate-pulse rounded-lg bg-suldak-gray-300" />
           </div>
         ))}
       </div>
